Clean up pickle vectorizer object code
Memory issues appears to be fixed. The root cause appears to be the version that created the pickled vectorizer is not necessarily the same as the vectorizer library version. Runs for over 30+ interactions, variety of search terms, with a variety of price ranges and dropdown selections on AWS web server.